Enable 'honor keyboard-interactive' in the global configuration
It's not possible to use default field's value as a default for a
checkbox so use a 'default' attribute.

Support for login check when configuring SCP site added by extending
SCPRepositoryPublisher.DescriptorImpl's doLoginCheck(StaplerRequest)
with a parameter. It's SCPRepositoryPublisher.DescriptorImpl's
doLoginCheck(StaplerRequest) method responsibility to convert
parameter values from string to boolean.

Signed-off-by: Mykola Nikishov <mn@mn.com.ua>